// src/lib/generated/SmartMailbox.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Predicate } from "./Predicate";

/**
 * A stored smart mailbox.
 */
export type SmartMailbox = { id: number, name: string, icon: string | null, predicate: Predicate, sortOrder: number, };